# src/docs/swaggerConfig.yaml
openapi: 3.0.0
info:
  title: Members-Only API
  description: A Sample API mimicking a social media platform, it allows members to access only authorized resources. .
  version: 1.0.0

  contact:
    name: Lekan Adetunmbi
    email: me@pollaroid.net
    url: https://wwww.pollaroid.net

  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api-mbo.herokuapp.com/v1
    description: Main (production) server
components:
  schemas:
    User:
      type: "object"
      properties:
        _id:
          type: "string"
          format: "uuid"
        name:
          type: "string"
        email:
          type: "string"
          format: "email"
        password:
          type: "string"
          format: "password"
        avatar:
          type: "string"
        lastLogin:
          type: "string"
          format: "date-time"
        isAdmin:
          type: "boolean"
        isMember:
          type: "boolean"
        resetPassword:
          type: "object"
          properties:
            code:
              type: "string"
            expiresBy:
              type: "string"
              format: "date-time"
        refreshToken:
          type: "object"
          properties:
            token:
              type: "string"
              format: "byte"
            expiresBy:
              type: "string"
              format: "date-time"
        tokenVersion:
          type: "integer"
          format: "int32"
        createdAt:
          type: "string"
          format: "date-time"
        updatedAt:
          type: "string"
          format: "date-time"
...
    Post:
      type: "object"
      properties:
        _id:
          type: "string"
          format: "uuid"
        user:
          type: "string"
          format: "uuid"
        postTitle:
          type: "string"
        postContent:
          type: "string"
        datePosted:
          type: "string"
          format: "date-time"
        comments:
          type: "array"
          items:
            type: "object"
            properties:
              comment_user:
                type: "string"
                format: "uuid"
              comment_list:
                type: "array"
                items:
                  type: "object"
                  properties:
                    comment:
                      type: "string"
                    comment_date:
                      type: "string"
                      format: "date-time"
  securitySchemes:
    BearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
    cookieAuth:
      type: "apiKey"
      in: "cookie"
      name: "jit"
  responses:
    400BadRequest:
      description: Bad request
    401Unauthorized:
      description: "Unauthorized"
...
    404NotFound:
      description: "The specified resource was not found"
...
    500ServerError:
      description: "Internal Server Error"
paths:
...
  /auth/refresh-token:
    post:
      tags:
        - Auth
      summary: Refresh Token
      description: Refreshes the current user's token
      operationId: refreshToken
      security:
        - cookieAuth: []
      responses:
        '200':
          description: 'Successfully refreshed token'
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  message:
                    type: "string"
                  token:
                    type: "string"
        '403':
          description: 'Refresh Token Validation Failures'
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  message:
                    type: "string"
        '404':
          $ref: '#/components/responses/404NotFound'
        '500':
          $ref: '#/components/responses/500ServerError'
...
  /password/reset-password:
    put:
      tags:
        - Password
      summary: Reset Password
      description: Resets the user's password
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: "object"
              properties:
                email:
                  type: "string"
                new_password:
                  type: "string"
                  format: "password"
                verification_code:
                  type: "string"
      responses:
        '200':
          description: "Success, Password Reset"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  message:
                    type: "string"
                  user:
                    $ref: "#/components/schemas/User"
        '400':
          $ref: '#/components/responses/400BadRequest'
        '403':
          description: "Verification Code Validation Failures"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  message:
                    type: "string"
        '404':
          $ref: '#/components/responses/404NotFound'
        '500':
          $ref: '#/components/responses/500ServerError'
...
  /posts/{id}:
    get:
      tags:
        - Post
      summary: Get Post By ID
      description: Returns the post with the given ID
      operationId: getPostById
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: "string"
      responses:
        '200':
          description: "Success, Post"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '404':
          $ref: '#/components/responses/404NotFound'
        '500':
          $ref: '#/components/responses/500ServerError'
...